Fixed cael's very smart solution for my fire addition stuff
