+++ src/main/java/com/nekomaster1000/infernalexp/mixin/client/MixinOverlayRenderer.java
@@ -16,7 +16,7 @@ public class MixinOverlayRenderer {
 
 	@ModifyVariable(method = "renderFire", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/client/renderer/model/RenderMaterial;getSprite()Lnet/minecraft/client/renderer/texture/TextureAtlasSprite;", ordinal = 0), name = "textureatlassprite")
 	private static TextureAtlasSprite IE_renderCustomFiresOverlay(TextureAtlasSprite original, Minecraft minecraftIn, MatrixStack matrixStackIn) {
-		return ((FireTypeAccess) minecraftIn.player).getFireType().getAssociatedSprite1().getSprite();
+		return ((FireTypeAccess) minecraftIn.player).getFireType().getSupplier().get().getAssociatedSprite1().getSprite();
 	}
 
 }
