+++ src/main/java/com/nekomaster1000/infernalexp/mixin/common/MixinEntity.java
@@ -1,10 +1,6 @@
-package com.nekomaster1000.infernalexp.mixin.client;
+package com.nekomaster1000.infernalexp.mixin.common;
 
-import com.nekomaster1000.infernalexp.access.FireTypeAccess;
-import net.minecraft.entity.Entity;
-import net.minecraft.nbt.CompoundNBT;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.At.Shift;
@@ -12,21 +8,16 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import com.nekomaster1000.infernalexp.access.FireTypeAccess;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.nbt.CompoundNBT;
+
 @Mixin(Entity.class)
 public abstract class MixinEntity implements FireTypeAccess {
 
 	@Unique
 	private KnownFireTypes fireType = KnownFireTypes.FIRE;
-
-	@Inject(method = "baseTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;setFlag(IZ)V", ordinal = 0, shift = Shift.AFTER))
-	private void IE_removeCustomFires(CallbackInfo ci) {
-		if (!isBurning()) {
-			((FireTypeAccess) this).setFireType(KnownFireTypes.FIRE);
-		}
-	}
-
-	@Shadow
-	abstract boolean isBurning();
 
 	@Inject(method = "writeWithoutTypeId", at = @At(value = "INVOKE", target = "Lnet/minecraft/nbt/CompoundNBT;putShort(Ljava/lang/String;S)V", ordinal = 0, shift = Shift.AFTER))
 	private void IE_writeCustomFires(CompoundNBT tag, CallbackInfoReturnable<CompoundNBT> ci) {
